Change discount type to float64

+++ pkg/db/types.go
@@ -48,7 +48,7 @@ type Discount struct {
 	Id string
 
 	Source   string
-	Discount int
+	Discount float64
 
 	During pgtype.Tstzrange
 }
